FinalProject.md
The foundational ideas around this project had everything to do with perspective. I wanted to take the user through different life scenarios and show them how their lives are affected depending on their state of mind. In my opinion, we're not mindful enough of how much our lives are influenced by our mental states. This project was my way of enlightening the user on the power of perspective that flies under our radars.